Adding base figure function

+++ src/settings.py
@@ -224,11 +224,11 @@ if SCENARIO_NAME == 'FragmentationKaandorp' and SUBMISSION in ['simulation', 'an
     # INITIAL PARTICLE SIZE (m)
     INIT_SIZE = int(os.environ['PARTICLE_SIZE']) * SIZE_FACTOR
     # INITIAL DENSITY (KG/M^3): 920 = polypropylene
-    INIT_DENSITY = float(os.environ['INIT_DENSITY'])
+    INIT_DENSITY = int(os.environ['INIT_DENSITY'])
 
 if SCENARIO_NAME == 'SizeTransport' and SUBMISSION in ['simulation', 'analysis']:
     # INITIAL DENSITY (KG/M^3): 920 = POLYPROPYLENE, 980 = HIGH DENSITY POLYETHYLENE (BRIGNAC ET AL. 2017)
-    INIT_DENSITY = float(os.environ['INIT_DENSITY'])
+    INIT_DENSITY = int(os.environ['INIT_DENSITY'])
 
 if SUBMISSION in ['simulation', 'analysis']:
     # FRAGMENTATION PROBABILITY
